@@ -24,19 +24,21 @@
     "postinstall": "npm rebuild node-sass"
   },
   "dependencies": {
+    "@angular-devkit/architect": "^0.1402.11",
     "@angular-devkit/build-optimizer": "^0.1202.18",
-    "@angular-devkit/build-webpack": "~0.1303.11",
-    "@angular-devkit/core": "^13.3.11",
-    "@angular-devkit/schematics": "^13.3.11",
-    "@angular/animations": "^13.4.0",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
+    "@angular-devkit/build-webpack": "^0.1402.11",
+    "@angular-devkit/core": "^14.2.11",
+    "@angular-devkit/schematics": "^14.2.11",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
     "@ngtools/webpack": "^11.2.3",
     "@schematics/angular": "^9.1.15",
     "@schematics/update": "^0.900.7",
@@ -56,16 +58,14 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1303.11",
-    "@angular-devkit/build-angular": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.11",
     "@angular-eslint/builder": "1.2.0",
     "@angular-eslint/eslint-plugin": "1.2.0",
     "@angular-eslint/eslint-plugin-template": "1.2.0",
     "@angular-eslint/schematics": "1.2.0",
     "@angular-eslint/template-parser": "1.2.0",
-    "@angular/cli": "13.3.11",
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/language-service": "13.4.0",
+    "@angular/cli": "14.2.11",
+    "@angular/language-service": "14.3.0",
     "@compodoc/compodoc": "^1.1.19",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
@@ -77,7 +77,7 @@
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jsdoc": "30.7.6",
     "eslint-plugin-prefer-arrow": "1.2.2",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "sass-lint": "1.13.1",
     "ts-node": "8.6.2",
     "typescript": "^4.6.4"
@@ -85,6 +85,9 @@
   "keywords": [
     "angular",
     "angular11",
+    "angular12",
+    "angular13",
+    "angular14",
     "material",
     "keyboard",
     "onscreen",
